pyproject.toml
[project]
name = "teams-ai-planner"
version = "0.1.0"
description = "Daily action planner that connects to Microsoft Teams, analyzes channel messages via AI, and suggests follow-up actions."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Workflow Copilot" }
]
dependencies = [
    "msal>=1.29.0",
    "openai>=1.53.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.2",
    "requests>=2.32.3",
    "rich>=13.9.2"
]

[project.urls]
Homepage = "https://github.com/example/teams-ai-planner"

[project.scripts]
teams-ai-planner = "teams_ai_planner.cli:main"
